split up tui tests
